+++ frontend/src/components/Routes/SHOP/Comp/Category/CategoryStyle.js
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const CategoriesStyled=styled.div`
+    background-color: white;
+`;
