+++ frontend/src/components/Routes/SHOP/Comp/Descriptioner/DescriptionerStyled.js
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const DescriptionerStyled=styled.div`
+    img{
+        width: min-content;
+    }
+`;
