+++ frontend/src/components/Routes/SHOP/Comp/Filter/FilterStyled.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const FilterStyled=styled.div`
+    width: fit-content;
+
+`;
