+++ frontend/src/components/Routes/SHOP/Comp/Paginner/PaginnerStyled.js
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const PaginnerStyled=styled.div``;
