+++ frontend/src/components/Routes/SHOP/Comp/Sorter/SorterStyled.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const SorterStyled=styled.div`
+    display: flex;
+    align-content: space-between;
+`;
