+++ frontend/src/components/Routes/SHOP/Containers/BigCont/BigContStyled.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const BigContStyled=styled.div`
+    width: 68%;
+    margin-left: 2%;
+    margin-right: 2%;
+
+`;
