+++ frontend/src/components/Routes/SHOP/Containers/Filterbar/Filterbar.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Categories from "../Comp/Categories";
-import Filter from "../Comp/Filter";
-import { FilterbarStyled } from "../styles";
+import Categories from "../../Comp/Category/Categories";
+import Filter from "../../Comp/Filter/Filter";
+import { FilterbarStyled } from "./FilterbarStyled";
 
 export default function Filterbar(){
     return (
