+++ frontend/src/components/Routes/SHOP/Containers/Filterbar/FilterbarStyled.js
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const FilterbarStyled=styled.div`
+    width: 28%
+`;
