+++ frontend/src/components/Routes/SHOP/Containers/ViewCards/ViewCardStyled.js
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const ViewcardStyled=styled.div``;
