+++ frontend/src/components/Routes/SHOP/Containers/ViewCards/ViewCards.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import ProductCard from "../../../common/Productcard";
-import { ViewcardStyled } from "../styles";
+import ProductCard from "../../../../common/Productcard/index";
+import { ViewcardStyled } from "./ViewCardStyled";
 
 export default function Viewcards(){
     const products=useSelector((state)=>state.products.isLoading?[]:state.products.products.products) || []
