+++ frontend/src/components/Routes/SHOP/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { useDispatch } from "react-redux"
 import { fetchAllCategories, fetchAllProducts } from "../../../redux/reducers/products/actions";
-import BigCont from "./Containers/BigCont";
-import Filterbar from "./Containers/Filterbar";
+import BigCont from "./Containers/BigCont/BigCont";
+import Filterbar from "./Containers/Filterbar/Filterbar";
 import { ShopStyled } from "./styles";
 
 export default function Shop(){
