+++ frontend/src/components/Routes/SHOP/styles.js
@@ -5,34 +5,3 @@ export const ShopStyled=styled.div`
     margin-top: 230px;
     background-color: #f7f7f7;
 `;
-
-export const CategoriesStyled=styled.div`
-    background-color: white;
-`;
-
-export const DescriptionerStyled=styled.div`
-    img{
-        width: min-content;
-    }
-`;
-export const FilterStyled=styled.div`
-    width: fit-content;
-
-`;
-export const PaginnerStyled=styled.div``;
-export const SorterStyled=styled.div`
-    display: flex;
-    align-content: space-between;
-`;
-
-export const BigContStyled=styled.div`
-    width: 68%;
-    margin-left: 2%;
-    margin-right: 2%;
-
-`;
-export const FilterbarStyled=styled.div`
-    width: 28%
-`;
-
-export const ViewcardStyled=styled.div``;
