+++ frontend/src/redux/reducers/products/actions.js
@@ -40,7 +40,7 @@ export const fetchAllProducts = function(){
   return async function (dispatch){
     dispatch(loading())
     try{
-      const res= await axios.get('http://localhost:3001/products');
+      const res= await axios.get('/products');
       dispatch(actionCreator(GET_ALL_PRODUCTS,res.data))
     }catch(error){
       console.log(error)
